Register Logic files added
To allow a user to register, the data must be collected from the
register form by the registerPost.php file and the data is then put into JSON   format to be sent to the server. At the moment the data is not being sent to the server successfully.
The FormUtils file has been added to allow error checks to be completed,
these are not yet implemented.